On Win32 run autodoc.pl and pod/perlmodlib.PL from the top level.
On *nix and VMS autodoc.pl and pod/perlmodlib.PL are both run from the top
level of the build tree. Previously Win32 built both with win32/ as the
current directory.

In win32/makefile.mk, move the invocation of $(PL2BAT) after the invocations
for autodoc.pl and pod/perlmodlib.PL to make the order consistent with that
of win32/Makefile.