Restore installman's -S option to some semblance of silence.
Commit a274383458745101 (in Oct 2001) converted installman to directly using
Pod::Man (instead of shelling out to pod2man), but added a print statement
which unconditionally output each manpage name. This print should be silenced
if -S (or --silence) is passed.

With this change installman -S generates no output on STDOUT. However, it
still generates about 75 lines of "no documentation in ..." warnings on
STDERR. As these are not fatal, arguably -S should silence these too.