Use U64TYPE instead of U64. The latter is a typedef only enabled for PERL_CORE
This isn't ideal. It would be nice to unconditionally "enable" U64, but it
might break anything which embeds perl but expects to declare its own U64.

It also isn't idea to assume that guessing U64TYPE will work. However, it
seems slightly better to attempt to use uint64_t if we have no U64TYPE, as
we might just not fail, whereas quitting immediately with an error is
guaranteed not to work.