No need to use UINT32_C() in the "Super-Fast" hash implementation.
The "Super-Fast" hash implementation had a definition of its get16bits()
macro which used UINT32_C(). UINT32_C() is part of C99 and isn't available
everywhere, particularly some C++ installations. It isn't actually needed
here, as the left shift 8 of a char value isn't going to overflow. So remove
it.