Data-Dumper: Boost test coverage and refactor.
Thanks to coverage data posted by Paul Johnson as part of Perl
Foundation-sponsored work, we know that as of Jan 01 2013,
Data::Dumper (v2.139) had the following coverage of its code by its
test suite:

                          stmt  bran  cond   sub
Dumper.xs                 94.8  64.5
blib/lib/Data/Dumper.pm   85.3  76.7  69.6  71.4

The coverage levels reported for Dumper.pm are unacceptably low,
particularly for a library which is part of the Perl 5 core
distribution.  With this commit, we begin to significantly improve
coverage.  (This branch/commit focuses on Dumper.pm, not Dumper.xs.)

Attempting to write tests for every line in a codebase frequently
uncovers places in the code that are superfluous or cannot be reached.
The close study of the code required to achieve high test coverage
also discloses places where the code and its documentation are at odds
with each other.  This work on Data::Dumper is no exception, so we
have made changes to Dumper.pm where appropriate.  In addition, in
order to facilitate this analysis and to make the codebase easier to
maintain going forward, the code in Dumper.pm has been tidied up in
certain locations, mostly with respect to tabs vs. whitespace in
indentation and cuddled elses.

Specifics of this commit:

* Modifications to Dumper.pm

** Eliminate subs Dumpf() and Dumpp, which were neither documented nor
tested anywhere in the codebase.  (They will be reinstated if and only
if someone can provide a rationale for that, along with documentation
and testing.)

** For both Values() and Names():  explicit handling (croak) of the
case where the method is provided with a defined argument that is not
an array reference.

** Tidying, mostly in _dump().

** Deletion of commented-out code.

* New test files

** t/dumpperl.t - test all branches of, and modes of triggering,
Dumpperl()

** t/indent.t - test Indent()

** t/misc.t - test mostly previously untested methods; more to come

** t/names.t - test Names()

** t/values.t - test Values()

** t/lib/Testing.pm - package holding subs used in the new test files

* Expanded test files

** t/terse.t - Test Terse()

The work done so far has significantly improved the coverage.  As of
Jan 05 2013 we are at:

                                  stmt  bran  cond   sub
Jan 01: blib/lib/Data/Dumper.pm   85.3  76.7  69.6  71.4
Jan 05: blib/lib/Data/Dumper.pm   92.4  83.3  77.4  84.6

More to come.