regcomp.c: Make inversion lists SVt_PVLV
This is the 2nd step in separating the inversion list body from header.
This commit gives inversion lists the header from a SVt_PVLV, and
repurposes one of its fields into being the length of the inversion
list.

This SV type was chosen because it has a sufficient number of fields to
accommodate all the header fields from inversion lists.