regcomp.c: Move some #defines to only file that uses them
These were used in a header file to provide synchronization between
files.  However, the only other file that would need them is a .pl file which
doesn't have access to them.  So simplify things so that the variables
are either removed entirely if only used in a single place, or
are #defined in the area where they are used