File::Spec in XS
For work purposes, last year I reimplemented parts of File::Spec in XS.
It wasn't in a CPANable form initially, and File::Spec had core/CPAN
version skew issues, so I didn't initially publish my work.  I've now
cleaned it up and rebased it on PathTools-3.40, and the patch is attached.
This is for review, and for whoever is currently maintaining File::Spec
to merge it into blead at an appropriate juncture.  (I fear it's a bit
too late to get into 5.18.)

[With a bump to a dev version on the changed .pm files to keep
t/porting/cmp_version.t sweet]