In the Makefile, depend on lib/buildcustomize.pl instead of using $(Icwd)
The Makefile macro Icwd was used to add appropriate includes for loading
the not-yet build Cwd early in the bootstrap. However, miniperl is now
automatically adding the relevant paths to @INC once lib/buildcustomize.pl
is available. So remove the use of $(Icwd), and instead depend on
lib/buildcustomize.pl

This removes duplication of the list of paths needed to add to @INC to be
able to load Cwd from dist. This information is already known to
write_buildcustomize.pl