In the Win32 makefiles, depend on buildcustomize.pl instead of using $(ICWD)
This ports the previous Makefile.SH change to the Win32 makefiles. However,
we can't eliminate $(ICWD) completely as it's also being used for some
invocations of the (real) perl binary. Only miniperl loads buildcustomize.pl
to set up @INC to include paths for the initial locations of modules such
as Cwd.