Fix regex /il and /iaa failures for single element [] class
This was a regression introduced in the v5.17 series.  It only affected
UTF-8 encoded patterns.  Basically, the code here should have
corresponded to, and didn't, similar logic located after the defchar:
label in this file.

We don't fold code points 0-255 under locale, as those aren't known
until run time.  Similarly, we don't allow folds that cross the 255/256
boundary, as those aren't well-defined; and under /aa we don't allow
folds that cross the 127/128 boundary.